% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_map_b37.R
\name{gwa_get_map}
\alias{gwa_get_map}
\title{Get genetic build 37 recombination map}
\usage{
gwa_get_map(pop = "CEU", write_map = FALSE)
}
\arguments{
\item{pop}{A \href{https://www.internationalgenome.org/category/population/}{1000 Genomes population code}. Default "CEU".}

\item{write_map}{Write a serialized dataframe of the map data to disk. Default \code{FALSE}, does not write the map to disk but instead returns a dataframe.}
}
\value{
A dataframe. If \code{write_map} is \code{TRUE}, then the recombination map is written to the working directory as genetic_map_b37_<\code{pop}>.rds. Note. the map is ~22MB.
}
\description{
Retrieves \href{https://www.internationalgenome.org/category/population/}{1000 Genomes phase 3 recombination map} data for a specified population.
}
\details{
Previously retrieved map data from https://mathgen.stats.ox.ac.uk/impute/1000GP_Phase3.tgz
}
