% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{gwa_region}
\alias{gwa_region}
\title{Regional plot}
\usage{
gwa_region(
  data,
  recomb_map,
  target,
  chromosome,
  target_bp,
  window_kb = 150,
  regulation = FALSE
)
}
\arguments{
\item{data}{A genetic association results dataframe.}

\item{recomb_map}{A recombination map dataframe with header and obligatory columns `position` (genomic position), `combined_rate` (combined recombination rate cM/Mb), and `chr` (chromosome). Use \code{\link{gwa_get_map}} to retrieve a build 37 map for any 1000 genomes phase 3 population.}

\item{target}{A character vector for the variant of interest.}

\item{chromosome}{An integer indicating the chromosome on which the region sits.}

\item{target_bp}{A integer indicating base pair position in the centre of the region.}

\item{window_kb}{A integer indicating window size in kb (default is 150kb), to be included either side of the \code{target_bp} location.}

\item{regulation}{A boolean (default \code{FALSE}) indicating whether or not to include regulation tracks.}
}
\description{
Plots is a negative log base 10 p-value against genomic position scatterplot for a relatively small genomic "window". Below this \bold{Association} layer, there are additional layers or "tracks" providing biological context: hg19 build 37 combined \bold{Recombination Rate}, known \bold{refSeq Genes}, and optional tracks including genetic regulation information.
}
\details{
You will need a genetic map. Use \code{\link{gwa_get_map}} to download 1000 Genomes phase 3 recombination map data.
}
