% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{A "manhattan" plot for genome-wide genetic association results}
\usage{
manhattan(
  data,
  trait_label = "phenotype",
  trim = 0.001,
  chromosome = "#CHROM",
  snp = "SNP",
  bp = "POS",
  a1 = "A1",
  effect = "BETA",
  p = "P",
  chr_color = c("#878D92", "#49494D"),
  strip_color = "grey25",
  ...
)
}
\arguments{
\item{data}{A genetic association results dataframe.}

\item{trait_label}{A character vector to be used as label. Default =
"phenotype".}

\item{trim}{Observations with p-values larger than the trim value
are dropped. Default = 0.001.}

\item{chromosome}{An integer vector of chromosome codes.}

\item{snp}{SNP id.}

\item{bp}{Base-pair coordinate.}

\item{a1}{Allele 1 (usually minor). See details.}

\item{effect}{Regression coefficient or odds ratio.}

\item{p}{Association p-value.}

\item{chr_color}{A character vector of 2 colours to be used on
alternate chromosomes.}

\item{strip_color}{Color of the vertical strip describing the panel.}

\item{...}{Additional arguments passed to \code{geom_point}.}
}
\description{
A scatter plot of negative log base 10 p-value against genomic
position
}
\details{
SNP id (\code{snp}), Allele 1 (\code{a1}), and regression
coefficient or odds ratio (\code{effect}) columns are only used for
labelling points. \code{a1} is described as it is in
\href{https://www.cog-genomics.org/plink/1.9/formats#assoc_linear}{PLINK}
documentation, but as this column is only used for labelling it can
refer to any allele. PLINK output names are used as the default
values for all columns.
}
